requirements.txt
streamlit
requests
beautifulsoup4
langchain
langchain-community
langchain-chroma
langchain-ollama
chromadb
ollama
pdfplumber
python-docx
odfpy